Remove non-bitmap tracking
